+++ helm/kyverno-policies-observability/templates/Silence.yaml
