--- helm/policies-aws/templates/CoreCAPI.yaml
@@ -1,25 +1,0 @@
-# THIS FILE IS GENERATED WITH 'make generate' - DO NOT EDIT MANUALLY
-apiVersion: kyverno.io/v1
-kind: ClusterPolicy
-metadata:
-  name: aws-core-capi-policy
-spec:
-  background: false
-  rules:
-  - name: kubeadmcontrolplane-disks
-    match:
-      resources:
-        kinds:
-        - controlplane.cluster.x-k8s.io/v1alpha3/KubeadmControlPlane
-        - controlplane.cluster.x-k8s.io/v1alpha4/KubeadmControlPlane
-        selector:
-          matchLabels:
-            cluster.x-k8s.io/watch-filter: capi
-    mutate:
-      patchesJson6902: |-
-        - path: "/spec/kubeadmConfigSpec/files/-"
-          op: add
-          value: { "content": "IyEvYmluL3NoCiMgY2hlY2sgZm9yIE5WTUUgZGlza3MKaWYgWyAiYGxzYmxrIHwgZ3JlcCBudm1lIHwgd2MgLWxgIiAtZ2UgNCBdOyB0aGVuCglFVENEX0RJU0s9Ii9kZXYvbnZtZTFuMSIKCUNPTlRBSU5FUkRfRElTSz0iL2Rldi9udm1lMm4xIgoJS1VCRUxFVF9ESVNLPSIvZGV2L252bWUzbjEiCgllY2hvIGZvdW5kIG52bWUgRElTS3MKZWxzZQoJRVRDRF9ESVNLPSIvZGV2L3h2ZGMiCiAgICAgICAgQ09OVEFJTkVSRF9ESVNLPSIvZGV2L3h2ZGQiCiAgICAgICAgS1VCRUxFVF9ESVNLPSIvZGV2L3h2ZGUiCmZpCgojIGluaXQgZXRjZCBkaXNrCm1rZnMueGZzICRFVENEX0RJU0sKbWtkaXIgLXAgL3Zhci9saWIvZXRjZAptb3VudCAkRVRDRF9ESVNLIC92YXIvbGliL2V0Y2QKCiMgaW5pdCBrdWJlbGV0IGRpc2sKbWtmcy54ZnMgJEtVQkVMRVRfRElTSwpta2RpciAtcCAvdmFyL2xpYi9rdWJlbGV0Cm1vdW50ICRLVUJFTEVUX0RJU0sgL3Zhci9saWIva3ViZWxldAoKIyBpbml0IGNvbnRhaW5lcmQgZGlzawpta2ZzLnhmcyAkQ09OVEFJTkVSRF9ESVNLCnN5c3RlbWN0bCBzdG9wIGNvbnRhaW5lcmQKcm0gLXJmIC92YXIvbGliL2NvbnRhaW5lcmQKbWtkaXIgLXAgL3Zhci9saWIvY29udGFpbmVyZAptb3VudCAkQ09OVEFJTkVSRF9ESVNLIC92YXIvbGliL2NvbnRhaW5lcmQKc3lzdGVtY3RsIHN0YXJ0IGNvbnRhaW5lcmQK", "encoding": "base64", "owner": "root", "path": "/opt/init-disks.sh" }
-        - path: "/spec/kubeadmConfigSpec/preKubeadmCommands/-1"
-          op: add
-          value: "/bin/sh /opt/init-disks.sh"
