--- helm/policies-aws/values.yaml
@@ -1,37 +1,0 @@
-region: "eu-west-1"
-sshKeyName: "ssh-key"
-
-awsClusterRoleIdentity:
-  roleARN: "default-arn"
-  sourceIdentityRef:
-    name: default
-    kind: AWSClusterControllerIdentity
-
-awsControlPlane:
-  instanceType: t3.large
-
-awsMachineTemplate:
-  rootVolume:
-    size: 120
-    type: gp3
-  nonRootVolumes:
-    etcd:
-      name: "/dev/xvdc"
-      encrypted: true
-      size: 100
-      type: gp3
-    varlibcontainerd:
-      name: "/dev/xvdd"
-      encrypted: true
-      size: 100
-      type: gp3
-    varlibkubelet:
-      name: "/dev/xvde"
-      encrypted: true
-      size: 100
-      type: gp3
-
-awsMachinePool:
-  rootVolume:
-    size: 300
-    type: gp3
