--- helm/policies-azure/values.yaml
@@ -1,1 +1,0 @@
-location: "westeurope"
