--- helm/policies-common/templates/Pod.yaml
@@ -1,74 +1,0 @@
-# THIS FILE IS GENERATED WITH 'make generate' - DO NOT EDIT MANUALLY
-{{- with .Values.Installation.V1 }}
-{{- if .Proxy.Enabled }}
-apiVersion: kyverno.io/v1
-kind: ClusterPolicy
-metadata:
-  name: inject-proxy-env
-spec:
-  rules:
-    - name: inject-proxy-env-to-containers
-      match:
-        resources:
-          kinds:
-            - Pod
-      mutate:
-        patchStrategicMerge:
-          spec:
-            containers:
-              - (name): "*"
-                env:
-                  {{- if .Secret.Proxy.HTTP }}
-                  - name: HTTP_PROXY
-                    value: {{ .Secret.Proxy.HTTP }}
-                  - name: http_proxy
-                    value: {{ .Secret.Proxy.HTTP }}
-                  {{- end }}
-                  {{- if .Secret.Proxy.HTTPS }}
-                  - name: HTTPS_PROXY
-                    value: {{ .Secret.Proxy.HTTPS }}
-                  - name: https_proxy
-                    value: {{ .Secret.Proxy.HTTPS }}
-                  {{- end }}
-                  {{- if .Proxy.NoProxy }}
-                  - name: NO_PROXY
-                    value: {{ join "," .Proxy.NoProxy }}
-                  - name: no_proxy
-                    value: {{ join "," .Proxy.NoProxy }}
-                  {{- end }}
-    - name: inject-proxy-env-to-init-containers
-      match:
-        resources:
-          kinds:
-            - Pod
-      preconditions:
-        any:
-        - key: "{{ `{{` }}request.object.spec.initContainers{{ `}}` }}"
-          operator: NotEquals
-          value: []
-      mutate:
-        patchStrategicMerge:
-          spec:
-            initContainers:
-              - (name): "*"
-                env:
-                  {{- if .Secret.Proxy.HTTP }}
-                  - name: HTTP_PROXY
-                    value: {{ .Secret.Proxy.HTTP }}
-                  - name: http_proxy
-                    value: {{ .Secret.Proxy.HTTP }}
-                  {{- end }}
-                  {{- if .Secret.Proxy.HTTPS }}
-                  - name: HTTPS_PROXY
-                    value: {{ .Secret.Proxy.HTTPS }}
-                  - name: https_proxy
-                    value: {{ .Secret.Proxy.HTTPS }}
-                  {{- end }}
-                  {{- if .Proxy.NoProxy }}
-                  - name: NO_PROXY
-                    value: {{ join "," .Proxy.NoProxy }}
-                  - name: no_proxy
-                    value: {{ join "," .Proxy.NoProxy }}
-                  {{- end }}
-{{- end }}
-{{- end }}
