--- helm/policies-common/templates/kyverno-policies-data.yaml
@@ -1,8 +1,0 @@
-# THIS FILE IS GENERATED WITH 'make generate' - DO NOT EDIT MANUALLY
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: kyverno-policies-data
-  namespace: default
-data:
-  nodeip: "{{ `{{` }} ds.meta_data.local_ipv4 {{ `}}` }}"
