--- helm/policies-common/values.yaml
@@ -1,10 +1,0 @@
-Installation:
-  V1:
-    Proxy:
-      Enabled: false
-    Guest:
-      Kubernetes:
-        Kubelet:
-          ImagePullProgressDeadline: 10m
-
-clusterDescription: "Unnamed Cluster"
