--- helm/policies-openstack/tests/ats/Pipfile
@@ -1,20 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pre-commit = "==2.11.1"
-setuptools = "==54.0.0"
-
-[packages]
-pytest-kube = {git = "https://github.com/giantswarm/pytest-kube.git", editable = false, ref = "kubectl"}
-pytest = "==6.2.2"
-pykube-ng = "==21.3.0"
-pyyaml = "==5.4.1"
-pytest-rerunfailures = "==9.1.1"
-requests = "==2.25.1"
-setuptools = "==54.0.0"
-
-[requires]
-python_version = "3.8"
