--- helm/policies-openstack/values.yaml
@@ -1,1 +1,0 @@
-failureDomain: ""
