--- helm/policies-shared/Chart.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-name: policies-shared
-appVersion: master
-description: Giant Swarm policies applied on all installations.
-home: https://github.com/giantswarm/kyverno-policies
-annotations:
-  application.giantswarm.io/team: "honeybadger"
-  config.giantswarm.io/version: 1.x.x
-version: [[ .Version ]]
