--- helm/policies-shared/templates/certconfig.yaml
@@ -1,35 +1,0 @@
-# THIS FILE IS GENERATED WITH 'make generate' - DO NOT EDIT MANUALLY
-apiVersion: kyverno.io/v1
-kind: ClusterPolicy
-metadata:
-  name: block-bulk-certconfigs-delete
-  annotations:
-    policies.kyverno.io/title: Block bulk certconfig deletes
-    policies.kyverno.io/subject: CertConfigs
-    policies.kyverno.io/description: >-
-      A bug in an old kubectl-gs version causes all certconfigs to
-      be deleted on performing a login, this policy blocks that
-      action while still allowing the single delete.
-spec:
-  validationFailureAction: enforce
-  background: false
-  rules:
-  - name: block-bulk-certconfigs-delete
-    match:
-      any:
-      - resources:
-          kinds:
-          - CertConfig
-    preconditions:
-      any:
-      - key: "{{ `{{` }}request.name{{ `}}` }}"
-        operator: Equals
-        value: ""
-    validate:
-      message: "Your current kubectl-gs version contains a critical bug, please update to the latest version using `kubectl gs selfupdate`"
-      deny:
-        conditions:
-        - key: "{{ `{{` }}request.operation{{ `}}` }}"
-          operator: In
-          value:
-          - DELETE
