--- helm/policies-shared/values.yaml
@@ -1,5 +1,0 @@
-managementCluster:
-  name: ""
-
-provider:
-  kind: ""
