--- helm/policies-vsphere/tests/ats/test_vsphere_default.py
@@ -1,29 +1,0 @@
-import sys
-sys.path.append('../../../tests')
-
-import yaml
-from functools import partial
-import time
-import random
-import string
-import ensure
-from textwrap import dedent
-
-from ensure import release
-
-import pytest
-from pytest_kube import forward_requests, wait_for_rollout, app_template
-
-import logging
-LOGGER = logging.getLogger(__name__)
-
-
-@pytest.mark.smoke
-def test_cluster_vsphere_policy(release) -> None:
-    """
-    test_cluster_policy tests defaulting of a Cluster where all required values are empty strings.
-
-    :param release: Release CR which is used by the Cluster.
-    """
-    assert release['metadata']['name'].startswith("v20.0.0")
-
