--- policies/common/kyverno-policies-data.yaml
@@ -1,7 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: kyverno-policies-data
-  namespace: default
-data:
-  nodeip: "{{ ds.meta_data.local_ipv4 }}"
