--- policies/openstack/CoreCAPI.yaml
@@ -1,19 +1,0 @@
-apiVersion: kyverno.io/v1
-kind: ClusterPolicy
-metadata:
-  name: capo-policy
-spec:
-  background: false
-  rules:
-  - name: machine-set-failure-domain
-    match:
-      resources:
-        kinds:
-        - cluster.x-k8s.io/v1alpha4/MachineSet
-        - cluster.x-k8s.io/v1beta1/MachineSet
-    mutate:
-      patchStrategicMerge:
-        spec:
-          template:
-            spec:
-              +(failureDomain): "{{ request.object.spec.template.metadata.labels.\"machine-deployment.giantswarm.io/failure-domain\" }}"
