--- policies/vsphere/VSphereCAPI.yaml
@@ -1,30 +1,0 @@
-apiVersion: kyverno.io/v1
-kind: ClusterPolicy
-metadata:
-  name: vsphere-capi-policy
-spec:
-  rules:
-  - name: vsphere-label
-    match:
-      resources:
-        kinds:
-        - infrastructure.cluster.x-k8s.io/v1alpha3/VSphereCluster
-        - infrastructure.cluster.x-k8s.io/v1alpha3/VSphereMachine
-        - infrastructure.cluster.x-k8s.io/v1alpha3/VSphereMachineTemplate
-        - infrastructure.cluster.x-k8s.io/v1alpha3/VSphereVM
-    context:
-    - name: releaseVersion
-      apiCall:
-        urlPath: "/apis/cluster.x-k8s.io/v1alpha3/namespaces/{{ request.object.metadata.namespace }}/clusters"
-        jmesPath: "items[?metadata.name == '{{request.object.metadata.labels.\"cluster.x-k8s.io/cluster-name\"}}'].metadata.labels.\"release.giantswarm.io/version\" | [0]"
-    preconditions:
-      any:
-      - key: "{{ releaseVersion }}"
-        operator: Equals
-        value: "20.0.0*"
-    mutate:
-      patchStrategicMerge:
-        metadata:
-          labels:
-            +(cluster.x-k8s.io/watch-filter): "capi"
-            cluster.x-k8s.io/watch-filter: "capi"
